requirements.txt
streamlit
opencv-python
mediapipe
diffusers
torch
torchvision
Pillow
numpy